[Bugfix][Diffusion] MiniMax-H3: promote Triton row offsets to int64 for long packed sequences
ref2va 768p/15s packs ~226k rows (reference video latents included). With the
qkv row stride of 10752, `token * stride` in _rms_norm_rope_kernel exceeded
INT32_MAX (226623 * 10752 ≈ 2.44e9) and wrapped to a negative offset, causing
an illegal memory access on the first denoise step (TP2, both ranks). The four
indexed modulation kernels in minimax_h3_modulation.py carry the same latent
overflow (row * hidden_size ≈ 1.62e9 today; only ~25% headroom before 1080p /
longer clips overflow).

Fix: index program_id(0) with .to(tl.int64) in all five kernels. Verified by:
- offline repro at the exact real shape (226624x28x128) crashed pre-fix, passes post-fix
- compute-sanitizer memcheck: 0 errors
- tests/diffusion/layers/test_fused_qk_norm_rope.py: 3 passed
- end-to-end 2xH800 TP2 ref2va 768p/15s: full 49-step denoise + valid 1344x768 output

repro_rope_oob.py is kept as the regression harness for this shape family.

+++ vllm_omni/diffusion/attention/ops/minimax_h3_modulation.py
@@ -39,7 +39,7 @@ def _indexed_scale_shift_kernel(
     row_offset,
     block_n: tl.constexpr,
 ):
-    row = tl.program_id(0) + row_offset
+    row = tl.program_id(0).to(tl.int64) + row_offset
     columns = tl.arange(0, block_n)
     mask = columns < hidden_size
     index = tl.load(indices_ptr + row * stride_indices)
@@ -71,7 +71,7 @@ def _indexed_gate_kernel(
     row_offset,
     block_n: tl.constexpr,
 ):
-    row = tl.program_id(0) + row_offset
+    row = tl.program_id(0).to(tl.int64) + row_offset
     columns = tl.arange(0, block_n)
     mask = columns < hidden_size
     index = tl.load(indices_ptr + row * stride_indices)
@@ -104,7 +104,7 @@ def _rms_norm_indexed_scale_shift_kernel(
     row_offset,
     block_n: tl.constexpr,
 ):
-    row = tl.program_id(0) + row_offset
+    row = tl.program_id(0).to(tl.int64) + row_offset
     columns = tl.arange(0, block_n)
     mask = columns < hidden_size
     index = tl.load(indices_ptr + row * stride_indices)
@@ -145,7 +145,7 @@ def _indexed_gate_rms_norm_scale_shift_kernel(
     row_offset,
     block_n: tl.constexpr,
 ):
-    row = tl.program_id(0) + row_offset
+    row = tl.program_id(0).to(tl.int64) + row_offset
     columns = tl.arange(0, block_n)
     mask = columns < hidden_size
     index = tl.load(indices_ptr + row * stride_indices)
